refactor(context): rename profileutil and add flag completions
