+++ pkg/cmd/context/contextcmdutil/flagset.go
@@ -23,6 +23,10 @@ func NewFlagSet(cmd *cobra.Command, f *factory.Factory) *FlagSet {
 
 // AddContextName adds a flag for setting the name of the context
 func (fs *FlagSet) AddContextName(name *string) *flagutil.FlagOptions {
+
+	svcContext, _ := fs.factory.ServiceContext.Load()
+
+	svcContextsMap := svcContext.Contexts
 	flagName := "name"
 
 	fs.StringVar(
@@ -31,6 +35,15 @@ func (fs *FlagSet) AddContextName(name *string) *flagutil.FlagOptions {
 		"",
 		fs.factory.Localizer.MustLocalize("context.common.flag.name"),
 	)
+
+	contextNames := make([]string, 0, len(svcContextsMap))
+	for k := range svcContextsMap {
+		contextNames = append(contextNames, k)
+	}
+
+	_ = fs.cmd.RegisterFlagCompletionFunc(flagName, func(cmd *cobra.Command, _ []string, _ string) ([]string, cobra.ShellCompDirective) {
+		return contextNames, cobra.ShellCompDirectiveNoSpace
+	})
 
 	return flagutil.WithFlagOptions(fs.cmd, flagName)
