+++ pkg/cmd/kafka/acl/create/create.go
@@ -10,8 +10,8 @@ import (
 	"github.com/redhat-developer/app-services-cli/pkg/core/ioutil/spinner"
 	"github.com/redhat-developer/app-services-cli/pkg/core/localize"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/connection"
+	"github.com/redhat-developer/app-services-cli/pkg/shared/contextutil"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/factory"
-	"github.com/redhat-developer/app-services-cli/pkg/shared/profileutil"
 	kafkainstanceclient "github.com/redhat-developer/app-services-sdk-go/kafkainstance/apiv1internal/client"
 	"github.com/spf13/cobra"
 )
@@ -237,7 +237,7 @@ func validateAndSetOpts(opts *aclcmdutil.CrudOptions) error {
 			return err
 		}
 
-		profileHandler := &profileutil.ContextHandler{
+		profileHandler := &contextutil.ContextHandler{
 			Context:   svcContext,
 			Localizer: opts.Localizer,
 		}
