+++ pkg/cmd/kafka/consumergroup/delete/delete.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 
 	kafkaflagutil "github.com/redhat-developer/app-services-cli/pkg/cmd/kafka/flagutil"
+	"github.com/redhat-developer/app-services-cli/pkg/shared/contextutil"
 	kafkacmdutil "github.com/redhat-developer/app-services-cli/pkg/shared/kafkautil"
-	"github.com/redhat-developer/app-services-cli/pkg/shared/profileutil"
 
 	"github.com/AlecAivazis/survey/v2"
 	"github.com/redhat-developer/app-services-cli/pkg/core/ioutil/iostreams"
@@ -58,7 +58,7 @@ func NewDeleteConsumerGroupCommand(f *factory.Factory) *cobra.Command {
 				return err
 			}
 
-			profileHandler := &profileutil.ContextHandler{
+			profileHandler := &contextutil.ContextHandler{
 				Context:   svcContext,
 				Localizer: opts.localizer,
 			}
