+++ pkg/cmd/kafka/consumergroup/resetoffset/reset_offset.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/redhat-developer/app-services-cli/pkg/cmd/kafka/consumergroup/groupcmdutil"
 	kafkaflagutil "github.com/redhat-developer/app-services-cli/pkg/cmd/kafka/flagutil"
+	"github.com/redhat-developer/app-services-cli/pkg/shared/contextutil"
 	kafkacmdutil "github.com/redhat-developer/app-services-cli/pkg/shared/kafkautil"
-	"github.com/redhat-developer/app-services-cli/pkg/shared/profileutil"
 
 	"github.com/AlecAivazis/survey/v2"
 	"github.com/redhat-developer/app-services-cli/pkg/core/cmdutil/flagutil"
@@ -82,7 +82,7 @@ func NewResetOffsetConsumerGroupCommand(f *factory.Factory) *cobra.Command {
 				return err
 			}
 
-			profileHandler := &profileutil.ContextHandler{
+			profileHandler := &contextutil.ContextHandler{
 				Context:   svcContext,
 				Localizer: opts.localizer,
 			}
