+++ pkg/cmd/registry/list/list.go
@@ -13,8 +13,8 @@ import (
 	"github.com/redhat-developer/app-services-cli/pkg/core/logging"
 	"github.com/redhat-developer/app-services-cli/pkg/core/servicecontext"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/connection"
+	"github.com/redhat-developer/app-services-cli/pkg/shared/contextutil"
 	"github.com/redhat-developer/app-services-cli/pkg/shared/factory"
-	"github.com/redhat-developer/app-services-cli/pkg/shared/profileutil"
 	srsmgmtv1 "github.com/redhat-developer/app-services-sdk-go/registrymgmt/apiv1/client"
 
 	"github.com/spf13/cobra"
@@ -122,7 +122,7 @@ func runList(opts *options) error {
 			return err
 		}
 
-		profileHandler := &profileutil.ContextHandler{
+		profileHandler := &contextutil.ContextHandler{
 			Context:   svcContext,
 			Localizer: opts.localizer,
 		}
