+++ pkg/shared/contextutil/util.go
@@ -1,4 +1,4 @@
-package profileutil
+package contextutil
 
 import (
 	"context"
